+++ cli/commands/asset/update.go
@@ -12,11 +12,11 @@ import (
 // UpdateCommand adds command that allows user to create new assets
 func UpdateCommand(cli *cli.SensuCli) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:          "update NAME",
+		Use:          "update [NAME]",
 		Short:        "update assets",
 		SilenceUsage: false,
 		RunE: func(cmd *cobra.Command, args []string) error {
-			// Print ot usage if we do not receive one argument
+			// Print out usage if we do not receive one argument
 			if len(args) != 1 {
 				return cmd.Help()
 			}