+++ cli/commands/configure/configure.go
@@ -27,6 +27,10 @@ func Command(cli *cli.SensuCli) *cobra.Command {
 		Short:        "Initialize sensuctl configuration",
 		SilenceUsage: true,
 		RunE: func(cmd *cobra.Command, args []string) error {
+			if len(args) != 0 {
+				return cmd.Help()
+			}
+
 			flags := cmd.Flags()
 
 			nonInteractive, err := flags.GetBool("non-interactive")
@@ -39,13 +43,13 @@ func Command(cli *cli.SensuCli) *cobra.Command {
 			if nonInteractive {
 				answers.withFlags(flags)
 			} else {
-				if err := answers.administerQuestionnaire(cli.Config); err != nil {
+				if err = answers.administerQuestionnaire(cli.Config); err != nil {
 					return err
 				}
 			}
 
 			// Write new API URL to disk
-			if err := cli.Config.SaveAPIUrl(answers.URL); err != nil {
+			if err = cli.Config.SaveAPIUrl(answers.URL); err != nil {
 				fmt.Fprintln(cmd.OutOrStderr())
 				return fmt.Errorf(
 					"unable to write new configuration file with error: %s",