+++ cli/commands/environment/delete.go
@@ -17,8 +17,7 @@ func DeleteCommand(cli *cli.SensuCli) *cobra.Command {
 		RunE: func(cmd *cobra.Command, args []string) error {
 			// If no name is present print out usage
 			if len(args) != 1 || args[0] == "" {
-				_ = cmd.Help()
-				return nil
+				return cmd.Help()
 			}
 
 			org := cli.Config.Organization()