+++ cli/commands/environment/update.go
@@ -10,14 +10,13 @@ import (
 // UpdateCommand allows the user to update environment
 func UpdateCommand(cli *cli.SensuCli) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:          "update NAME",
+		Use:          "update [NAME]",
 		Short:        "update environment description",
 		SilenceUsage: false,
 		RunE: func(cmd *cobra.Command, args []string) error {
-			// Print ot usage if we do not receive one argument
+			// Print out usage if we do not receive one argument
 			if len(args) != 1 {
-				_ = cmd.Help()
-				return nil
+				return cmd.Help()
 			}
 
 			// Fetch environment from API