+++ cli/commands/event/resolve.go
@@ -1,7 +1,6 @@
 package event
 
 import (
-	"errors"
 	"fmt"
 
 	"github.com/sensu/sensu-go/cli"
@@ -16,8 +15,7 @@ func ResolveCommand(cli *cli.SensuCli) *cobra.Command {
 		SilenceUsage: true,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			if len(args) != 2 {
-				_ = cmd.Help()
-				return errors.New("missing argument(s)")
+				return cmd.Help()
 			}
 
 			entity := args[0]