+++ cli/commands/event/show.go
@@ -22,8 +22,7 @@ func ShowCommand(cli *cli.SensuCli) *cobra.Command {
 		SilenceUsage: true,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			if len(args) != 2 {
-				_ = cmd.Help()
-				return nil
+				return cmd.Help()
 			}
 
 			// Fetch event from API