+++ cli/commands/role/create.go
@@ -1,7 +1,6 @@
 package role
 
 import (
-	"errors"
 	"fmt"
 
 	"github.com/sensu/sensu-go/cli"
@@ -12,12 +11,12 @@ import (
 // CreateCommand defines new command to create roles
 func CreateCommand(cli *cli.SensuCli) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:          "create NAME",
+		Use:          "create [NAME]",
 		Short:        "create new roles",
 		SilenceUsage: true,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			if len(args) != 1 {
-				return errors.New("must provide role's name as first argument")
+				return cmd.Help()
 			}
 
 			role := &types.Role{Name: args[0]}