+++ cli/commands/user/add_role.go
@@ -1,7 +1,6 @@
 package user
 
 import (
-	"errors"
 	"fmt"
 
 	"github.com/sensu/sensu-go/cli"
@@ -11,14 +10,13 @@ import (
 // AddRoleCommand adds a command that allows admin's to add a role to a user
 func AddRoleCommand(cli *cli.SensuCli) *cobra.Command {
 	return &cobra.Command{
-		Use:          "add-role USERNAME ROLE",
+		Use:          "add-role [USERNAME] [ROLE]",
 		Short:        "add role to user given username and role",
 		SilenceUsage: true,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			// If no name is present print out usage
 			if len(args) != 2 {
-				_ = cmd.Help()
-				return errors.New("bad arguments")
+				return cmd.Help()
 			}
 
 			username := args[0]