+++ cli/commands/user/remove_role.go
@@ -10,14 +10,13 @@ import (
 // RemoveRoleCommand adds a command that allows admin's to remove a role from a user
 func RemoveRoleCommand(cli *cli.SensuCli) *cobra.Command {
 	return &cobra.Command{
-		Use:          "remove-role USERNAME ROLE",
+		Use:          "remove-role [USERNAME] [ROLE]",
 		Short:        "remove role from user given username and role",
 		SilenceUsage: true,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			// If no name is present print out usage
 			if len(args) != 2 {
-				_ = cmd.Help()
-				return nil
+				return cmd.Help()
 			}
 
 			username := args[0]