+++ CHANGELOG.md
@@ -7,15 +7,21 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 ### Added
-- Add .gitattributes file with merge stratety for the Changelog.
+- Add .gitattributes file with merge strategy for the Changelog.
 - Context switcher added for dashboard
 - Add api support for adhoc check requests.
+- Added better error checking for CLI commands and support for mutually
+exclusive fields.
+- Added `--interactive` flag to CLI which is required to run interactive mode.
 
 ### Changed
 - Silenced `begin` supports human readable time (Format: Jan 02 2006 3:04PM MST)
 in `sensuctl` with optional timezone. Stores the field as unix epoch time.
 - Increased the timeout in the store's watchers tests
 - Incremental retry mechanism when waiting for agent and backend in e2e tests.
+
+### Fixed
+ - Fixed required flags in `sensuctl` so requirements are enforced.
 
 ## [2.0.0-alpha.16] - 2018-02-07
 ### Added